Updates builtin quickfix action.
This takes advantage of the new ability of denite to reopen itself on an
action (before the behavior did not work, and was undefined. See
https://github.com/Shougo/denite.nvim/issues/387)
